+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 <!-- ### Security 🛡️-->
 
 ## [Unreleased]
+## [5.13.0] - 2025-09-11
+### Added ✔️
+* **OntimizeJdbcDaoSupport**: Created executeSQLStatement() to use DDL statements. #175
+## [5.12.1] - 2025-04-11
+### Changed 🛠️
+* **OntimizeJdbcDaoSupport**: Changed visibility of composeSQL method (protected → public) #174
 ## [5.12.0] - 2025-03-12
 ### Added ✔️
 * **Multi tenant**: Allow to configure tenants using the preferences. #153
@@ -61,7 +67,9 @@
 * **POM**: Sorted pom alphabetically and sorted, extracted version to properties and put all dependencies into dependency manager.
 * **Sonar**: Fix some sonar code smells.
 
-[unreleased]: https://github.com/ontimize/ontimize-jee/compare/5.12.0...HEAD
+[unreleased]: https://github.com/ontimize/ontimize-jee/compare/5.13.0...HEAD
+[5.13.0]: https://github.com/ontimize/ontimize-jee/compare/5.12.1...5.13.0
+[5.12.1]: https://github.com/ontimize/ontimize-jee/compare/5.12.0...5.12.1
 [5.12.0]: https://github.com/ontimize/ontimize-jee/compare/5.11.0...5.12.0
 [5.11.0]: https://github.com/ontimize/ontimize-jee/compare/5.10.0...5.11.0
 [5.10.0]: https://github.com/ontimize/ontimize-jee/compare/5.9.0...5.10.0